Removed the ServiceManager of the intermediate layer
