+++ src/base/cservices.cpp
@@ -3,11 +3,8 @@
 #include <memory>
 #include <thread>
 #include <unordered_map>
-
-
 #include <assert.h>
 #include "cservices-thread.h"
-#include "cservices.h"
 
 extern "C"
 {
@@ -23,6 +20,23 @@ bool hasLibeventThread = false;
 #else
     static void keepalive_timer_cb(uv_timer_t* handle) {}
 #endif
+    
+void globalInit(void(*postFunc)(void*, void*), uint32_t options, const char* logPath, size_t logSize)
+{
+    if (logPath)
+    {
+        karere::gLogger.logToFile(logPath, logSize);
+    }
+    services_init(postFunc, options);
+}
+    
+void globalCleanup()
+{
+#ifndef KARERE_DISABLE_WEBRTC
+    rtcModule::globalCleanup();
+#endif
+    services_shutdown();
+}
     
 MEGAIO_EXPORT eventloop* services_get_event_loop()
 {
@@ -96,75 +110,6 @@ MEGAIO_EXPORT int services_shutdown()
     hasLibeventThread = false;
     SVC_LOG_INFO("Libevent thread terminated");
     return 0;
-}
-
-//Handle store
-
-struct HandleItem
-{
-    unsigned short type;
-    void* ptr;
-    HandleItem(unsigned short aType, void* aPtr):type(aType), ptr(aPtr){}
-    bool operator==(const HandleItem& other) const
-    {
-        return ((type == other.type) && (ptr == other.ptr));
-    }
-};
-
-std::unordered_map<megaHandle, HandleItem> gHandleStore;
-megaHandle gHandleCtr = 0;
-
-MEGAIO_EXPORT void* services_hstore_get_handle(unsigned short type, megaHandle handle)
-{
-    auto it = gHandleStore.find(handle);
-    if ((it == gHandleStore.end()) || (it->second.type != type))
-        return nullptr;
-    return it->second.ptr;
-}
-
-MEGAIO_EXPORT megaHandle services_hstore_add_handle(unsigned short type, void* ptr)
-{
-#ifndef NDEBUG
-    megaHandle old = gHandleCtr;
-#endif
-    megaHandle id = ++gHandleCtr;
-#ifndef NDEBUG
-    if (id < old)
-    {
-        fprintf(stderr, "ERROR: megaHandle id generator wrapped around\n");
-        fflush(stderr);
-        abort();
-    }
-    bool inserted =
-#endif
-    gHandleStore.emplace(std::piecewise_construct,
-        std::forward_as_tuple(id), std::forward_as_tuple(type, ptr))
-#ifndef NDEBUG
-        .second
-#endif
-               ;
-    assert(inserted);
-    return id;
-}
-
-MEGAIO_EXPORT int services_hstore_remove_handle(unsigned short type, megaHandle handle)
-{
-    auto it = gHandleStore.find(handle);
-    if (it == gHandleStore.end())
-    {
-#ifndef NDEBUG
-        fprintf(stderr, "ERROR: services_hstore_remove_handle: Handle not found (id=%u, type=%d)\n", handle, type);
-#endif
-        return 0;
-    }
-    if (it->second.type != type)
-    {
-        fprintf(stderr, "ERROR: services_hstore_remove_handle: Handle found, but requested type %u does not match actual type %u\n", type, it->second.type);
-        fflush(stderr);
-        return 0;
-    }
-    gHandleStore.erase(it);
-    return 1;
 }
 
 /*int64_t services_get_time_ms()
