+++ src/megachatapi_impl.h
@@ -22,8 +22,6 @@
 #ifndef MEGACHATAPI_IMPL_H
 #define MEGACHATAPI_IMPL_H
 
-#define USE_LIBWEBSOCKETS 1
-
 #include "megachatapi.h"
 
 //the megaapi.h header needs this defined externally
@@ -654,26 +652,6 @@ public:
     void push_front(void *event);
     void* pop();
 };
-
-/**
- * @brief Any app using this library requires to init some services at startup.
- * They will be started along with the construction of the first MegaChatApi object created by the app.
- * When the app terminates, the resources required by the aforementioned services will be automatically
- * released. In case the app is terminated abruptly and you have a chance to explicitly release them,
- * call ServiceManager::cleanup.
- */
-class ServiceManager
-{
-public:
-    static void init();
-    static void cleanup();
-    ~ServiceManager();
-
-private:
-    ServiceManager();
-    static std::shared_ptr<ServiceManager> mInstance;
-};
-
 
 class MegaChatApiImpl :
         public karere::IApp,
